Selectors, Events and Templates working very good

+++ assets/js/jquery.js
@@ -9,7 +9,6 @@ let events = [
     "change",
     "load",
     "error",
-    "select",
     "submit",
     "option_change",
     "mouseover",
@@ -21,15 +20,31 @@ let eventBtn = $('<button class="events" id = "btn-events">EVENTS</button>');
 $(".q-links").append(eventBtn);
 eventBtn.on("click", () => {
     createEventNavBar();
-    readyEvent();
-    if (destroyFunctionNavBar) {
+    // readyEventjQuery();
+    eventConstructors(ready, "ev-0");
+    eventConstructors(click, "ev-1");
+    eventConstructors(dbClick, "ev-2");
+    eventConstructors(keyPress, "ev-3");
+    eventConstructors(mouseEvent, "ev-4");
+    eventConstructors(changeInput, "ev-5");
+    eventConstructors(loadImg, "ev-6");
+    eventConstructors(errorImg, "ev-7");
+    eventConstructors(submitted, "ev-8");
+    eventConstructors(optionChange, "ev-9");
+    eventConstructors(mouseOver, "ev-10");
+    eventConstructors(changeCheck, "ev-11");
+    eventConstructors(clickedElement, "ev-12");
+
+    if (destroyFunctionNavBar && destroySelectorNavBar) {
         destroyFunctionNavBar();
+        destroySelectorNavBar();
     }
 });
 
 /*create event nav bar*/
 
 function createEventNavBar() {
+    $("#nav-bar").html("");
     for (let i = 0; i < events.length; i++) {
         let navBar = $('<button class = "event-nav-bar nav-btn"> </button>');
         // navBar.attr("href", "#" + events[i]);
@@ -64,16 +79,6 @@ let functions = [
     "hide()",
     "show()",
     "fadeIn()_fadeOut()",
-    "each().css()",
-    "parent()",
-    "children()",
-    "getByClass",
-    "getById",
-    "getAllByClass",
-    "attr_selected",
-    "first()",
-    "input.value",
-    "remove_all_selectors",
     "animate()",
 ];
 
@@ -83,14 +88,34 @@ let functionBtn = $(
 $(".q-links").append(functionBtn);
 functionBtn.on("click", () => {
     createFunctionNavBar();
-    if (destroyEventNavBar) {
+    eventConstructors(createHtml, "fc-0");
+    eventConstructors(removeHtml, "fc-1");
+    eventConstructors(appendC, "fc-2");
+    eventConstructors(prepend, "fc-3");
+    eventConstructors(insertAfter, "fc-4");
+    eventConstructors(insertBefore, "fc-5");
+    eventConstructors(cloning, "fc-6");
+    eventConstructors(addClass, "fc-7");
+    eventConstructors(removeClass, "fc-8");
+    eventConstructors(toggleClass, "fc-9");
+    eventConstructors(addDisabled, "fc-10");
+    eventConstructors(removeDisabled, "fc-11");
+    eventConstructors(setDataAttr, "fc-12");
+    eventConstructors(removeDataAttr, "fc-13");
+    eventConstructors(hideEle, "fc-14");
+    eventConstructors(showEle, "fc-15");
+    eventConstructors(fadeInOut, "fc-16");
+    eventConstructors(animaTe, "fc-17");
+    if (destroyEventNavBar && destroySelectorNavBar) {
         destroyEventNavBar();
+        destroySelectorNavBar();
     }
 });
 
 /*create function nav bar*/
 
 function createFunctionNavBar() {
+    $("#nav-bar").html("");
     for (let i = 0; i < functions.length; i++) {
         let navBar = $('<button class = "function-nav-bar nav-btn"> </button>');
         // navBar.attr("href", "#" + functions[i]);
@@ -105,51 +130,102 @@ function destroyFunctionNavBar() {
     $(".function-nav-bar").remove();
 }
 
+/*-------------------------------------------------------SELECTORS-----------------------------------------------------------------*/
+
+let selectors = [
+    "each().css()",
+    "parent()",
+    "children()",
+    "getByClass",
+    "getById",
+    "getAllByClass",
+    "attr_selected",
+    "first()",
+    "input.value",
+    "remove_all_selectors",
+];
+
+let selectorBtn = $(
+    '<button class="selectors" id = "btn-selectors">SELECTORS</button>'
+);
+$(".q-links").append(selectorBtn);
+selectorBtn.on("click", () => {
+    createSelectorNavBar();
+    eventConstructors(eachCss, "sc-0");
+    eventConstructors(PaRent, "sc-1");
+    eventConstructors(chiLd, "sc-2");
+    eventConstructors(getByKlass, "sc-3");
+    eventConstructors(getByYd, "sc-4");
+    eventConstructors(allKlass, "sc-5");
+    eventConstructors(attrSelect, "sc-6");
+    eventConstructors(primaSelect, "sc-7");
+    eventConstructors(inputValue, "sc-8");
+    eventConstructors(removeAllSe, "sc-9");
+
+    if (destroyEventNavBar && destroyFunctionNavBar) {
+        destroyEventNavBar();
+        destroyFunctionNavBar();
+    }
+});
+
+/*create function nav bar*/
+
+function createSelectorNavBar() {
+    $("#nav-bar").html("");
+    for (let i = 0; i < selectors.length; i++) {
+        let navBar = $('<button class = "selector-nav-bar nav-btn"> </button>');
+        // navBar.attr("href", "#" + selectors[i]);
+        navBar.attr("id", "sc-" + parseInt(i));
+        navBar.text(selectors[i]);
+        $("#nav-bar").append(navBar);
+        navBar.css({ display: "flex" }, { "flex-direction": "column" });
+    }
+}
+
+function destroySelectorNavBar() {
+    $(".selector-nav-bar").remove();
+}
+
 /*-------------------------------------------------------Activate templates-------------------------------------------------------*/
-let ready = `<template id="ready">
-      <h2>
-        When the HTML document has been loaded and you can manipulate it with
-        JavaScript
-      </h2>
-      <h3>jQuery</h3>
-      <code>
-        $(document).ready(() => { console.log('Document is ready!'); });
-      </code>
 
-      <h3>Vanilla</h3>
-      <code>
-        document.addEventListener("load", () => { console.log("Document is
-        ready!"); });
-      </code>
-      <h3>Sources:</h3>
-      <ul>
-      <li><a
-        href="https://www.techiediaries.com/javascript/document-ready-vs-window-onload-vs-window-load/"
-      >https://www.techiediaries.com/javascript/document-ready-vs-window-onload-vs-window-load/</a></li>
-      <li><a href="https://www.youtube.com/watch?v=OFKDCJIgxUk">https://www.youtube.com/watch?v=OFKDCJIgxUk</a></li>
-      <div class="vanilla">
+/*-------------------------------------------------------END OF NAV BAR-------------------------------------------------------*/
 
-      </div>
-      <div class="query">
+/*-------------------------------------------------------EVENTS-----------------------------------------------------------------*/
+// function readyEventjQuery() {
+//     $("#ev-0").on("click", () => {
+//         $(document).on("load", () => {
+//             alert("Document is jQ_ready!");
+//         });
+//     });
+// }
 
-      </div>
-      <script>
-          $(document).ready(() => { console.log('Document is ready!'); });
-          document.addEventListener("load", () => { console.log("Document is
-        ready!"); });
-      </script>
-    </template>`;
+// function activateEvent() {
+//     console.log("Document is jQ_ready!");
+//     if ($("#ev-1").length) {
 
-function readyEvent() {
-    let ele1 = document.getElementById("ev-0");
+//         $("#ev-1").on("click", () => {
+//             clickEventjQuery();
+//             clickEventJS();
+//         });
+//     }
+// }
+
+// function clickEventjQuery() {
+//     $(".demo").html = "";
+//     $(".question").html = "When an HTML item has been clicked";
+//     $("jQuery").on("click", () => {
+//             $(".jQuery").html = "jQuery";
+//         }
+//     }
+
+function eventConstructors(x, y) {
+    let ele1 = document.getElementById(y);
     ele1.addEventListener("click", () => {
-        console.log(ele1);
         let container = document.querySelector(".content");
-        container.insertAdjacentHTML("beforeend", ready);
+        container.innerHTML = "";
+        container.insertAdjacentHTML("beforeend", x);
         let readys = document.querySelector("#ready").content;
         let readysCopy = document.importNode(readys, true);
         container.appendChild(readysCopy);
     });
 }
-
-/*-------------------------------------------------------END OF NAV BAR-------------------------------------------------------*/
